--- myWebPage/client/src/hooks/useFetch.tsx
@@ -1,26 +1,0 @@
-import axios from "axios";
-import { useState, useEffect } from "react";
-
-export default function useFetch<T>(URL: string) {
-    const [data, setData] = useState<T | null>(null);
-    const [loading, setLoading] = useState(true);
-    const [error, setError] = useState<string | null>(null);
-
-    useEffect(() => {
-        const fetchData = async () => {
-            try {
-                const response = await axios.get<T>(URL);
-                setData(response.data);
-            } catch (err) {
-                setError(err instanceof Error ? err.message : "Unknown error");
-            } finally {
-                setLoading(false);
-            }
-        };
-
-        fetchData();
-    }, [URL]);
-
-    return { data, loading, error };
-}
-
